InferenceServer/inference/requirements.txt
opencv-python
autogluon.vision
flask
flask_cors
boto3
botocore
